Fixed implicit conversions

+++ src/main.cpp
@@ -30,20 +30,20 @@ static int run_unit_tests(int argc, const char* argv[]) {
    return result;
 }
 
-inline constexpr uint64_t pow10(uint32_t exponent) {
-   return (exponent == 0) ? 1 : (10ULL * pow10(exponent - 1));
+inline constexpr int64_t pow10(uint32_t exponent) {
+   return (exponent == 0) ? 1 : (10 * pow10(exponent - 1));
 }
 
 template <uint32_t DECIMAL_PLACES>
 struct fixed_point_t {
 
-   static constexpr uint64_t exponent = pow10(DECIMAL_PLACES);
+   static constexpr int64_t exponent = pow10(DECIMAL_PLACES);
 
    constexpr fixed_point_t()
       : m_value() {}
 
    constexpr fixed_point_t(double value)
-      : m_value(value * exponent) {}
+      : m_value(static_cast<int64_t>(value * exponent)) {}
 
    fixed_point_t& operator+=(const fixed_point_t& other) {
       m_value += other.m_value;
@@ -58,7 +58,6 @@ struct fixed_point_t {
    int64_t whole_number() const {
       return m_value / exponent;
    }
-
 
 private:
    int64_t m_value;
